Define co-inductive Diverges and a failed attempt of a proof
I attempted to prove:

  Diverges e <=> eval e = Diverge

Everything went smoothly up until the Proj case of the ==>
direction of the theorem. The problem there is that Proj can
pick from within a Constructor value a Diverge element.

An option could be to define all of eval as a co-inductive
relation Eval, and prove:

  Eval e v <=> eval e = v

Another option could be to define an alternative version of
eval that uses a value type where the Constructor contains
a list of expressions instead of values.