Drop Diverge clause from unfold_rel
This way, the construction corresponds more closely to the construction
from Pitts' chapter---divergence is a phenomenon present in his language
too, yet divergence receives no special treatment in similarity.

With this definition, divergence becomes a bottom element of the
simulation relation, i.e. everything simulates a divergent expression.

However, crucially, bisimilarity will still satisfy

  x ~ y /\ eval x = Diverge ==> eval y = Diverge

as a straightforward consequence of app_bisimilarity_similarity.

This resolves the issue from @myreen's previous commit by removing
the need for a divergence case when proving that the Howe
construction is a simulation relation.