Patch pure_to_thunk top-level proof
This commit adds a version of the pureLang itree semantics that inserts
Tick expressions at Bind applications, and proves it equivalent to the
regular pureLang semantics. The pure_to_thunk compiler is proven correct
wrt. the former semantics and the theorem is then turned into a theorem
about the real top-level semantics for pureLang.

Leaves a few straight-forward (but tedious) cheats in and around a
simulation proof stating that it is OK to introduce Tick expressions.