Case compilation algorithm in SML
This is from
https://julesjacobs.com/notes/patternmatching/patternmatching.pdf

It compiles arbitrarily complicated/nested patterns into ones where
constructors are only applied to variable arguments.  It keeps
underscore patterns, which can be compiled away when you have a
complete list of all constructors to hand.

This will guide the actual algorithm that gets written in HOL.